+++ src/utils/date.ts
@@ -1,0 +1,9 @@
+export function formatDate(d: Date): string {
+  const year = d.getFullYear();
+  const month = ("0" + (d.getMonth() + 1)).slice(-2);
+  const day = ("0" + d.getDate()).slice(-2);
+  const hour = ("0" + d.getHours()).slice(-2);
+  const minute = ("0" + d.getMinutes()).slice(-2);
+
+  return `${year}-${month}-${day} ${hour}:${minute}`;
+}
