+++ src/views/dataset/DatasetInfo.vue
@@ -73,16 +73,7 @@ import {
   IDisplayLayer,
   newLayer
 } from "../../store/model";
-
-function formatDate(d: Date): string {
-  const year = d.getFullYear();
-  const month = ("0" + (d.getMonth() + 1)).slice(-2);
-  const day = ("0" + d.getDate()).slice(-2);
-  const hour = ("0" + d.getHours()).slice(-2);
-  const minute = ("0" + d.getMinutes()).slice(-2);
-
-  return `${year}-${month}-${day} ${hour}:${minute}`;
-}
+import { formatDate } from "@/utils/date";
 
 @Component
 export default class DatasetInfo extends Vue {
